requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
PySocks==1.7.1
pytest==7.4.3
requests==2.31.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
